Add Poll action and BatteryPowered device type.
BatteryPowered devices have two events, "Heartbeat"
and "LowBattery". "Heartbeat" is sent periodically
to indicate that the device is still alive. "LowBattery"
is sent if a low battery is detected periodically or
when the device status changes.

The "Heartbeat" event requires the device be linked to
the modem as a responder on group 4. The "LowBattery"
event requires the same on group 3. Currently we don't
yet have anything to set that up via the WebThings Gateway.

The Poll action just gets the remote status and updates
device properties if there was a change. Battery
devices cannot be polled, so the action is not
available there.